requirements.txt
langchain==0.1.20
langchain-community
langgraph==0.0.21
tavily
python-dotenv
requests
rich
google-generativeai
langchain-google-genai